Factorisation du code, un tir de missile ne supprime plus les bonus
