+++ src/TirVaisseau/SMissileH.java
@@ -23,15 +23,6 @@ public class SMissileH extends SMissile{
         this.move(30, -30);
     }
 
-    @Override
-    public void effect(Objet objet) {
-        if(this.collision(objet)){
-            if(objet.isEnnemy()){
-                this.game().remove(objet);
-                this.game().remove(this);
-            }
-        }
-    }
